Collapse JdbcShardFailureTestCase and JdbcShardFailureIT
